+++ src/index.ts
@@ -5,6 +5,10 @@ export * from "./components/Text/index.js";
 export * from "./components/Header/index.js";
 export * from "./components/Footer/index.js";
 export * from "./components/Layout/index.js";
+export * from "./components/Avatar/index.js";
+export * from "./components/HamburgerMenu/index.js";
+export * from "./components/Menu/index.js";
+export * from "./components/Menu/MenuItem.js";
 
 /** Typed theme input for generating a CSS file at build time. */
 export type Theme = {